@@ -4,51 +4,54 @@
   "scripts": {
     "ng": "ng",
     "start": "ng serve",
-    "build": "ng build --configuration production"
+    "build": "ng build --configuration production",
+    "lint": "eslint --report-unused-disable-directives --max-warnings 0",
+    "postinstall": "patch-package"
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.3.3",
-    "@angular/cdk": "^17.3.3",
-    "@angular/common": "^17.3.3",
-    "@angular/compiler": "^17.3.3",
-    "@angular/core": "^17.3.3",
-    "@angular/forms": "^17.3.3",
-    "@angular/platform-browser": "^17.3.3",
-    "@angular/platform-browser-dynamic": "^17.3.3",
-    "@angular/router": "^17.3.3",
+    "@angular/animations": "^20.3.10",
+    "@angular/cdk": "^20.2.12",
+    "@angular/common": "^20.3.10",
+    "@angular/compiler": "^20.3.10",
+    "@angular/core": "^20.3.10",
+    "@angular/forms": "^20.3.10",
+    "@angular/platform-browser": "^20.3.10",
+    "@angular/platform-browser-dynamic": "^20.3.10",
+    "@angular/router": "^20.3.10",
     "@materia-ui/ngx-monaco-editor": "^6.0.0",
-    "angular-split": "^17.2.0",
+    "angular-split": "^20.0.0",
     "assert": "^2.1.0",
-    "fs-browser-stub": "^1.0.1",
-    "js-beautify": "^1.15.1",
+    "events": "^3.3.0",
+    "js-beautify": "^1.15.4",
     "knex": "^3.1.0",
-    "monaco-editor": "^0.44.0",
-    "mssql": "^10.0.2",
-    "rxjs": "~7.8.1",
-    "sql-formatter": "^15.3.0",
-    "tslib": "^2.6.2",
-    "url": "^0.11.3",
+    "monaco-editor": "^0.54.0",
+    "rxjs": "~7.8.2",
+    "sql-formatter": "^15.6.10",
+    "stream": "^0.0.3",
+    "tslib": "^2.8.1",
     "util": "^0.12.5",
-    "zone.js": "^0.14.4"
+    "zone.js": "^0.15.1"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.3.3",
-    "@angular/cli": "^17.3.3",
-    "@angular/compiler-cli": "^17.3.3",
+    "@angular/build": "^20.3.9",
+    "@angular/cli": "^20.3.9",
+    "@angular/compiler-cli": "^20.3.10",
+    "@eslint/compat": "^1.4.1",
     "@types/js-beautify": "^1.14.3",
-    "@types/node": "^20.12.4",
-    "prettier": "^3.2.5",
-    "ts-node": "~10.9.2",
-    "tslint": "^6.1.3",
-    "typescript": "^5.4.4"
+    "@types/node": "^24.10.0",
+    "angular-eslint": "^20.5.1",
+    "eslint": "^9.39.1",
+    "eslint-config-prettier": "^10.1.8",
+    "eslint-plugin-prettier": "^5.5.4",
+    "eslint-plugin-unicorn": "^62.0.0",
+    "patch-package": "^8.0.1",
+    "prettier": "^3.6.2",
+    "typescript": "^5.9.3",
+    "typescript-eslint": "^8.46.3"
   },
   "browser": {
-    "crypto": false,
-    "fs": "fs-browser-stub",
-    "os": false,
-    "path": false,
-    "stream": false,
+    "fs": false,
     "timers": false,
     "tty": false
   }
