@@ -1,13 +1,4 @@
-import { enableProdMode } from "@angular/core";
-import { platformBrowserDynamic } from "@angular/platform-browser-dynamic";
-
+import { platformBrowser } from "@angular/platform-browser";
 import { AppModule } from "./app/app.module";
-import { environment } from "./environments/environment";
 
-if (environment.production) {
-  enableProdMode();
-}
-
-platformBrowserDynamic()
-  .bootstrapModule(AppModule)
-  .catch(err => console.error(err));
+void platformBrowser().bootstrapModule(AppModule);
